Make log-level parameter optional
With the current behaviour as soon as we extend the base test case we cannot individually run the tests in the IDE.
Behind the scenes IDE creates a temporary testng.xml file every time we run the tests for a class or a single test.

Since log-level is mandatory now we cannot run individual tests in the IDE (Intellij) since the temporary testng.xml created by the IDE
doesn't contain this parameter. This fix makes the log-level parameter optional.